+++ tests/build.gradle
@@ -5,5 +5,7 @@ sourceSets.test.java.srcDirs = ["src/test/java"]
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:1.5.21"
-    implementation group: 'org.bouncycastle', name: 'bcprov-jdk15on', version: '1.68'
+    implementation group: 'org.bouncycastle', name: 'bcprov-jdk15on', version: '1.69'
+    implementation group: 'org.bouncycastle', name: 'bcpkix-jdk15on', version: '1.69'
+    implementation group: 'org.bouncycastle', name: 'bctls-jdk15on', version: '1.69'
 }
